+++ src/features/dashboard/presentation/pages/dashboard-page.tsx
@@ -61,18 +61,13 @@ export default function DashboardPage() {
     { date: "14/04", Clics: 62, Leads: 30, Conversions: 8 },
   ];
 
-  const [isNotificationPanelOpen, setIsNotificationPanelOpen] = useState(false)
-  const [notifications, setNotifications] = useState(allNotifications)
+  const [notifications, setNotifications] = useState(allNotifications);
 
   const handleDismissNotification = (id: number) => {
-    setNotifications((prev) => prev.filter((notification) => notification.id !== id))
-  }
-
-  const handleMarkAllRead = () => {
-    setNotifications((prev) => prev.map((notification) => ({ ...notification, read: true })))
-  }
-    // Get latest 3 unread notifications for dashboard display
-  const dashboardNotifications = notifications.filter((n) => !n.read).slice(0, 3)
+    setNotifications((prev) => prev.filter((notification) => notification.id !== id));
+  };
+  // Get latest 3 unread notifications for dashboard display
+  const dashboardNotifications = notifications.filter((n) => !n.read).slice(0, 3);
 
   return (
     <div className="space-y-6">
@@ -84,31 +79,26 @@ export default function DashboardPage() {
       </div>
 
       {dashboardNotifications.length > 0 && (
-            <div className="space-y-3">
-              <div className="flex items-center justify-between">
-                <h3 className="text-sm font-medium text-gray-900">Recent Notifications</h3>
-                <Button
-                  variant="ghost"
-                  size="sm"
-                  onClick={() => setIsNotificationPanelOpen(true)}
-                  className="text-sm text-muted-foreground hover:text-foreground"
-                >
-                  View More
-                  <ChevronRight className="h-3 w-3 ml-1" />
-                </Button>
-              </div>
+        <div className="space-y-3">
+          <div className="flex items-center justify-between">
+            <h3 className="text-sm font-medium text-gray-900">Notifications récentes</h3>
+            <Button variant="ghost" size="sm" className="text-sm text-muted-foreground hover:text-foreground">
+              Voir plus
+              <ChevronRight className="h-3 w-3 ml-1" />
+            </Button>
+          </div>
 
-              <div className="space-y-2">
-                {dashboardNotifications.map((notification) => (
-                  <NotificationCard
-                    key={notification.id}
-                    notification={notification}
-                    onDismiss={handleDismissNotification}
-                  />
-                ))}
-              </div>
-            </div>
-          )}
+          <div className="space-y-2">
+            {dashboardNotifications.map((notification) => (
+              <NotificationCard
+                key={notification.id}
+                notification={notification}
+                onDismiss={handleDismissNotification}
+              />
+            ))}
+          </div>
+        </div>
+      )}
 
       <div className="grid w-full gap-4 md:grid-cols-2 lg:grid-cols-4">
         <Card>
