+++ src/presentation/routes/app-routing.tsx
@@ -5,11 +5,11 @@ import { PageSkeleton } from "@/presentation/components/loaders/page-skeleton";
 import { AuthLayout } from "@/presentation/layouts/auth-layout";
 import { NotFound } from "@/presentation/components/not-found";
 import { SpaceLayout } from "@/presentation/layouts/space-layout";
-import CreateCampaign from "@/features/campaigns/presentation/create-new-campaigns";
 
 const LoginPage = lazy(() => import("@/features/auth/presentation/pages/login-page"));
 const DashboardPage = lazy(() => import("@/features/dashboard/presentation/pages/dashboard-page"));
 const CampaignListPage = lazy(() => import("@/features/campaigns/presentation/pages/campaign-list-page"));
+const CreateCampaignPage = lazy(() => import("@/features/campaigns/presentation/pages/create-campaign-page"));
 const LeadListPage = lazy(() => import("@/features/leads/presentation/pages/lead-list-page"));
 const TransactionListPage = lazy(() => import("@/features/transactions/presentation/pages/transaction-list-page"));
 const ProfilePage = lazy(() => import("@/features/profile/presentation/pages/profile-page"));
@@ -42,7 +42,7 @@ export const router = createBrowserRouter([
             path: "create",
             element: (
               <Suspense fallback={<PageSkeleton />}>
-                <CreateCampaign />
+                <CreateCampaignPage />
               </Suspense>
             ),
           },
